Add `Path`, `Paths` and `Plot` to public API
These types allow one to plot something consisting out of multiple
paths. A `Path` is build from one or more `Point`s where each point
is a coordinate in a 2D cartesian plane.

The plotter isn't aware of coordinates; you can't instruct the
plotter to move to position (-127, 123). Instead, the plotter moves
relative to it's current position.

The difference between two `Point`s is a `Vector`. And `Vector`s can be
translated into a series of `Command`s that control the plotters
movement.