Fix preview for non-continuous paths
The preview for a plot with multiple non-continuous was wrong. Every
path would start where the previous path ended.

For paths that doen't start at the end of the previous paths, that's
problematic. This problem has been fixed.

This commit introduces a `Coordinate`, which is generic type parameter.
`Path`s are build out of a collection of absolute `Coordinate`s.