Fix bug that rendered distinct paths as a continuous path
Until now, distinct path were rendered as continues paths.
A bug caused a path to start where the previous path ended.

When plotting a map with roads, this issue became really present.

This commit replaces Point and Vector with Coordinate<T>. T is either Absolute or Relative.

A Path is now a series of absolute coordinates. This Path is turned into a Stroke, which has an absolute two coordinates
indicating the start and end of the stroke. The coordinates between start and end are relative to each other.

After a Stroke has been plotted, the plotter head moves back to home, before moving to the start of the subsequent
Stroke.

Fixes #8